Add option to disable the delete button on items when multiple=true

+++ src/slektr.js
@@ -455,14 +455,18 @@ export default class Slektr {
       this.renderValueContent(badge, option);
       el.appendChild(badge);
 
-      let deleteButton = document.createElement('button');
-      deleteButton.className = 'slektr-multiple-option-delete';
-      el.appendChild(deleteButton);
 
-      deleteButton.slektr_value = value;
-      let deleteIcon = deleteIconElement();
-      deleteButton.appendChild(deleteIcon);
-      deleteButton.addEventListener("click", this.removeValueFromMultiple, true);
+      if (!this.config.disableDelete) {
+        let deleteButton = document.createElement('button');
+        deleteButton.className = 'slektr-multiple-option-delete';
+        el.appendChild(deleteButton);
+
+        deleteButton.slektr_value = value;
+        let deleteIcon = deleteIconElement();
+        deleteButton.appendChild(deleteIcon);
+        deleteButton.addEventListener("click", this.removeValueFromMultiple, true);
+      }
+
 
       this.fieldEl.appendChild(el);
     }
@@ -529,6 +533,8 @@ export default class Slektr {
   }
 
   buildOptionsFilter(container) {
+    if (this.config.hideSearch) return;
+
     let el = document.createElement('div');
     el.className = 'slektr-options-filter-container';
     container.append(el);
